Move SessionInfo to api
